refactor(renderer): streamline notification handling by updating notify settings structure and introducing a new notify DOM parser
