+++ packages/frameworks/angular/projects/renderer/src/lib/genui-renderer.ts
@@ -164,7 +164,7 @@ export class GenuiRenderer implements OnInit {
 
   updateContextAndState() {
     this.rendererSettingsService.syncFromParent();
-    this.instance?.setNotify(this.rendererSettingsService.settings);
+    this.instance?.setNotify(this.rendererSettingsService.settings.notify);
     this.instance?.setContext({
       callAction: this.callAction.bind(this),
     });
